==> 01_get_sp_meta_file.ipynb <==
import json
from core.connectors import Connector

# Load file json from  data/raw/site_file_urls.json
with open('./data/raw/songhong_file_urls.json') as file:
    data = json.load(file)

# Load processed IDs
with open('./data/raw/processed_urls.json') as file:
    loaded_processed_urls = json.load(file)

# Load all data
with open('./data/raw/all_data.json') as file:
    all_data = json.load(file)

# Connect to SharePoint
SP_Connector = Connector.to_sharepoint()

processed_urls = list(loaded_processed_urls.values())

for n, data_item in enumerate(list(data.values())[:2]):
    print(f"=====Process {n+1}/{len(data)} =====")
    if data_item in processed_urls:
        print("Already processed")
    else:
        try:
            file_meta = SP_Connector.get_metadata(data_item)
            file_id = file_meta['GUID']
            # Add to all_data
            all_data[file_id] = file_meta

        except Exception as e:
            print(f"Error: {e}")
            continue
        
        # Save dict to json file
        with open('./data/raw/all_data.json', 'w') as f:
            json.dump(all_data, f, indent=4) 

        # Save processed ids
        processed_urls.append(data_item)
        processed_urls_dict = { i : url for i, url in enumerate(processed_urls)}

        # Save dict to json file
        with open('./data/raw/processed_urls.json', 'w') as f:
            json.dump(processed_urls_dict, f, indent=4)

        print("Finished: ", data_item.split('/')[-1])
