==> 05_embeddings_values.ipynb <==
import os, json
import pandas as pd
from dotenv import load_dotenv; load_dotenv()
from open_ai import MyOAI
OAI = MyOAI(os.getenv("OPENAI_API_KEY"))

# Load file entities
file_path = 'data/processed/file_entities.csv'
saving_file_path = 'data/embeddings/emb_file_entities.csv'
files_data = pd.read_csv(file_path)
for i in range(len(files_data['Name'])):
    files_data['Name'][i] = files_data['Name'][i].replace("'", " ")
    files_data['ReportName'][i] = str(files_data['ReportName'][i]).replace("'", " ")
# Load Basin entities
basin_file_path = 'data/processed/basin_entities.csv'
block_file_path = 'data/processed/block_entities.csv'
well_file_path = 'data/processed/well_entities.csv'
author_file_path = 'data/processed/author_entities.csv'
contractor_file_path = 'data/processed/contractor_entities.csv'
type_of_format_file_path = 'data/processed/type_of_format_entities.csv'
types_of_report_file_path = 'data/processed/types_of_report_entities.csv'
status_of_report_file_path = 'data/processed/status_of_report_entities.csv'
fields_file_path = 'data/processed/fields_entities.csv'
function_file_path = 'data/processed/function_entities.csv'
type_of_file_file_path = 'data/processed/type_of_file_entities.csv'

# Saving path
saving_basin_file_path = 'data/embeddings/emb_basin_entities.csv'
saving_block_file_path = 'data/embeddings/emb_block_entities.csv'
saving_well_file_path = 'data/embeddings/emb_well_entities.csv'
saving_author_file_path = 'data/embeddings/emb_author_entities.csv'
saving_contractor_file_path = 'data/embeddings/emb_contractor_entities.csv'
saving_type_of_format_file_path = 'data/embeddings/emb_type_of_format_entities.csv'
saving_types_of_report_file_path = 'data/embeddings/emb_types_of_report_entities.csv'
saving_status_of_report_file_path = 'data/embeddings/emb_status_of_report_entities.csv'
saving_fields_file_path = 'data/embeddings/emb_fields_entities.csv'
saving_function_file_path = 'data/embeddings/emb_function_entities.csv'
saving_type_of_file_file_path = 'data/embeddings/emb_type_of_file_entities.csv'

basin_data = pd.read_csv(basin_file_path)
# Load Block entities
block_data = pd.read_csv(block_file_path)
# Load Well entities
well_data = pd.read_csv(well_file_path)
# Load Author entities
author_data = pd.read_csv(author_file_path)
# Load Contractor entities
contractor_data = pd.read_csv(contractor_file_path)
# Load TypeofFormat entities
type_of_format_data = pd.read_csv(type_of_format_file_path)
# Load TypesOfReport entities
types_of_report_data = pd.read_csv(types_of_report_file_path)
# Load StatusOfReport entities
status_of_report_data = pd.read_csv(status_of_report_file_path)
# Load Fields entities
fields_data = pd.read_csv(fields_file_path)
# Load Function entities
function_data = pd.read_csv(function_file_path)
# Load TypeOfFile entities
type_of_file_data = pd.read_csv(type_of_file_file_path)


### Embedding Function

def get_embeddings(data:pd.DataFrame, column_name:str, saving_path:str):
    # Remove nan values of ReportAuthor
    data = data[data[column_name].notna()]
    unique_data = data[column_name].unique()
    embeded_unique_data = {}
    for i, item in enumerate(unique_data):
        print(f"{i+1}/{len(unique_data)}")
        emb = OAI.get_embedding(item)
        embeded_unique_data[item] = emb
    data[column_name + "Embeddings"] = data[column_name].apply(lambda x: embeded_unique_data[x])
    data = data.reset_index(drop=True)
    # Save to csv
    data.to_csv(saving_path, index=False)
    return data

#### File Embeddings

emb_files_data = True
if not emb_files_data:
    files_data = files_data[files_data['Name'].notna()]
    unique_files_data = files_data['Name'].unique()
    # Create new column = value in Name + ReportName column
    files_data['EmbText'] = files_data['Name'] + ' ' + files_data['ReportName']
    # files_data['NameEmbeddings'] = files_data['Name'].apply(lambda x: OAI.get_embedding(x))
    name_embeddings = []
    for i, name in enumerate(files_data['Name']):
        print(f"{i+1}/{len(files_data['Name'])}")
        embedding = OAI.get_embedding(name)
        name_embeddings.append(embedding)

    files_data['NameEmbeddings'] = name_embeddings
    data = files_data.reset_index(drop=True)
    # Save to csv
    data.to_csv(saving_file_path, index=False)

#### Author Embeddings

embeddings = True
if not embeddings:
    emb_authors = get_embeddings(data=author_data, 
                column_name="ReportAuthor",
                saving_path=saving_author_file_path,
                )
    emb_authors

#### Contractors

contractor_embeddings = True
if not contractor_embeddings:
    emb_contractors = get_embeddings(data=contractor_data, 
                column_name="Contractors",
                saving_path=saving_contractor_file_path,
                )
    emb_contractors
