==> 06_update_emb_neo4j.ipynb <==
import pandas as pd
from neo4j import GraphDatabase

URI = "neo4j://localhost"
AUTH = ("neo4j", "167trungkinh")

# URI="neo4j+s://18429959.databases.neo4j.io"
# AUTH = ("neo4j","6hXdC44Aa0Tm9UD3s5xGve3itUUNH9afYOcFOp2Z3ic")

saving_file_path = 'data/embeddings/emb_file_entities.csv'
saving_author_file_path = 'data/embeddings/emb_author_entities.csv'
saving_contractor_file_path = 'data/embeddings/emb_contractor_entities.csv'
files_data = pd.read_csv(saving_file_path)
author_data = pd.read_csv(saving_author_file_path)
contractor_data = pd.read_csv(saving_contractor_file_path)
# Remove duplicate Author values
author_data = author_data.drop_duplicates(subset=['ReportAuthor'], keep='first')
author_data = author_data.reset_index(drop=True)
contractor_data = contractor_data.drop_duplicates(subset=['Contractors'], keep='first')
contractor_data = contractor_data.reset_index(drop=True)

with GraphDatabase.driver(URI, auth=AUTH) as driver:
    for i in range(len(files_data)):

        query = """
        MATCH (n:File {{name: '{}'}}) 
        SET n.embeddings = {}
        """.format(
                    files_data['Name'][i],
                    files_data['NameEmbeddings'][i],
                    )
        # print(query)
        print(f"{i+1}/{len(files_data)}")
        driver = GraphDatabase.driver(URI, auth=AUTH)
        driver.execute_query(query_=query, database_="neo4j")

with GraphDatabase.driver(URI, auth=AUTH) as driver:
    for i in range(len(author_data)):

        query = """
        MATCH (n:Author {{name: '{}'}}) 
        SET n.embeddings = {}
        """.format(
                    files_data['ReportAuthor'][i],
                    files_data['AuthorEmbedding'][i],
                    )
        # print(query)
        print(f"{i+1}/{len(files_data)}")
        driver = GraphDatabase.driver(URI, auth=AUTH)
        driver.execute_query(query_=query, database_="neo4j")

with GraphDatabase.driver(URI, auth=AUTH) as driver:
    for i in range(len(author_data)):

        query = """
        MATCH (n:Contractor {{name: '{}'}}) 
        SET n.embeddings = {}
        """.format(
                    files_data['Contractors'][i],
                    files_data['ContractorsEmbeddings'][i],
                    )
        # print(query)
        print(f"{i+1}/{len(files_data)}")
        driver = GraphDatabase.driver(URI, auth=AUTH)
        driver.execute_query(query_=query, database_="neo4j")
